Increased timeout to give more time for generating XML files.

+++ app/models/FileRetrival.scala
@@ -33,7 +33,7 @@ object FileRetrival {
          val authenticator: StaticUserAuthenticator = new StaticUserAuthenticator(ftpSite, userName, password)
          val opts: FileSystemOptions = new FileSystemOptions()
          DefaultFileSystemConfigBuilder.getInstance().setUserAuthenticator(opts, authenticator)
-         FtpFileSystemConfigBuilder.getInstance().setDataTimeout(opts, new Integer(600));
+         FtpFileSystemConfigBuilder.getInstance().setDataTimeout(opts, new Integer(10000));
          var fileObj = VFS.getManager().resolveFile("ftp://" + ftpSite + filePath, opts)
          val inputStream = fileObj.getContent().getInputStream()
          return Source.fromInputStream(inputStream).mkString
